+++ src/Components/Buttons/Actions/Restore.php
@@ -7,10 +7,10 @@ namespace BladeUIKitBootstrap\Components\Buttons\Actions;
 use BladeUIKitBootstrap\Components\Buttons\FormButton;
 use Illuminate\Support\Str;
 
-class Logout extends FormButton
+class Restore extends FormButton
 {
     public function __construct(
-        public string $action = '',
+        public string $action,
         public ?string $text = null,
         public ?string $formId = null,
         public ?string $title = null,
@@ -19,12 +19,11 @@ class Logout extends FormButton
         public bool $noOutline = false,
         public bool $disabled = false,
         public bool $novalidate = true,
-        string $method = 'POST',
-        string $variant = 'secondary',
+        string $method = 'PATCH',
+        string $variant = 'warning',
     ) {
-        $action = $action === '' ? \route('logout') : $action;
-        $text ??= trans('Logout');
-        $formId ??= 'logout-'.Str::random(32);
+        $text ??= ucfirst(trans('actions.restore'));
+        $formId ??= 'restore-'.Str::random(32);
 
         parent::__construct($action, $text, $formId, $title, $confirm, $outline, $noOutline, $disabled, $novalidate, $method, $variant);
     }