+++ src/Concerns/BtnType.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+namespace BladeUIKitBootstrap\Concerns;
+
+use InvalidArgumentException;
+
+trait BtnType
+{
+    public string $type = 'button';
+
+    private const ALLOWED_BUTTON_TYPE = [
+        'button',
+        'reset',
+        'submit',
+    ];
+
+    private function validType(string $type): string
+    {
+        if (! \in_array($type, self::ALLOWED_BUTTON_TYPE)) {
+            throw new InvalidArgumentException(\sprintf(
+                'The button type "%s" is not allowed. Allowed type are: %s.',
+                e($type),
+                \implode(', ', self::ALLOWED_BUTTON_TYPE)
+            ));
+        }
+
+        return $type;
+    }
+}